Split up tracing instance modules
Now that we have partitioned where we need the tracing instances in the
protocol construction we can also partition the defintion of the tracing
instances too, to make the whole lot a bit more managable.

We split the instances up into a few categories:
 * common
 * network
 * consensus
 * mock ledger
 * byron ledger

The big ones are the network and consensus, with the ledger ones being a
bit smaller. The old module was a whopper.

Again establishing this pattern will make adding another module for
Shelly a lot cleaner and more comprehensible.
