+++ cardano-config/cardano-config.cabal
@@ -26,7 +26,11 @@ library
                        Cardano.Config.Topology
                        Cardano.Config.TraceConfig
                        Cardano.Config.Types
-                       Cardano.Tracing.ToObjectOrphans
+                       Cardano.TracingInstances.Byron
+                       Cardano.TracingInstances.Common
+                       Cardano.TracingInstances.Consensus
+                       Cardano.TracingInstances.Mock
+                       Cardano.TracingInstances.Network
 
   build-depends:       base >=4.12 && <5
                      , aeson
