+++ cardano-config/src/Cardano/TracingInstances/Network.hs
@@ -1,0 +1,585 @@
+{-# LANGUAGE FlexibleContexts      #-}
+{-# LANGUAGE FlexibleInstances     #-}
+{-# LANGUAGE LambdaCase            #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE OverloadedStrings     #-}
+{-# LANGUAGE ScopedTypeVariables   #-}
+{-# LANGUAGE TypeFamilies          #-}
+{-# LANGUAGE UndecidableInstances  #-}
+{-# LANGUAGE NamedFieldPuns        #-}
+
+{-# OPTIONS_GHC -fno-warn-orphans  #-}
+{-# OPTIONS_GHC -Wno-all-missed-specialisations #-}
+
+module Cardano.TracingInstances.Network
+  ( showTip
+  , showPoint
+  ) where
+
+import           Cardano.Prelude hiding (show)
+import           Prelude (String, show, id)
+
+import           Data.Text (pack)
+
+import qualified Network.Socket as Socket (SockAddr)
+import           Network.Mux (WithMuxBearer (..), MuxTrace (..))
+
+import           Cardano.TracingInstances.Common
+
+import           Ouroboros.Network.Block
+import           Ouroboros.Network.BlockFetch.ClientState
+                   (TraceFetchClientState (..), TraceLabelPeer (..))
+import           Ouroboros.Network.BlockFetch.Decision
+                   (FetchDecision, FetchDecline (..))
+import           Ouroboros.Network.Codec (AnyMessage (..))
+import qualified Ouroboros.Network.NodeToClient as NtC
+import qualified Ouroboros.Network.NodeToNode   as NtN
+import           Ouroboros.Network.NodeToNode
+                   (WithAddr(..), ErrorPolicyTrace(..), TraceSendRecv (..))
+import           Ouroboros.Network.Protocol.BlockFetch.Type
+                   (BlockFetch, Message(..))
+import           Ouroboros.Network.Protocol.TxSubmission.Type
+                   (Message (..), TxSubmission)
+import           Ouroboros.Network.Snocket (LocalAddress (..))
+import           Ouroboros.Network.Subscription
+                   (ConnectResult (..), DnsTrace (..), SubscriptionTrace (..),
+                    SubscriberError (..), WithDomainName (..), WithIPList (..))
+import           Ouroboros.Network.TxSubmission.Inbound
+                   (TraceTxSubmissionInbound(..))
+import           Ouroboros.Network.TxSubmission.Outbound
+                   (TraceTxSubmissionOutbound (..))
+
+-- We do need some consensus imports to provide useful trace messages for some
+-- network protocols
+import           Ouroboros.Consensus.Util.Condense (Condense, condense)
+import           Ouroboros.Consensus.Mempool.API
+                   (GenTx, HasTxId, HasTxs(..), TxId, txId)
+
+
+showTip :: Condense (HeaderHash blk)
+        => TracingVerbosity
+        -> Tip blk
+        -> String
+showTip verb = showPoint verb . getTipPoint
+
+showPoint :: Condense (HeaderHash blk)
+          => TracingVerbosity
+          -> Point blk
+          -> String
+showPoint verb pt =
+  case pt of
+    GenesisPoint -> "genesis (origin)"
+    BlockPoint slot h -> trim (condense h) ++ "@" ++ condense slot
+ where
+  trim :: [a] -> [a]
+  trim = case verb of
+    MinimalVerbosity -> take 7
+    NormalVerbosity -> take 7
+    MaximalVerbosity -> id
+
+
+--
+-- * instances of @HasPrivacyAnnotation@ and @HasSeverityAnnotation@
+--
+-- NOTE: this list is sorted by the unqualified name of the outermost type.
+
+instance HasPrivacyAnnotation NtC.HandshakeTr
+instance HasSeverityAnnotation NtC.HandshakeTr where
+  getSeverityAnnotation _ = Info
+
+
+instance HasPrivacyAnnotation NtN.HandshakeTr
+instance HasSeverityAnnotation NtN.HandshakeTr where
+  getSeverityAnnotation _ = Info
+
+
+instance HasPrivacyAnnotation NtN.AcceptConnectionsPolicyTrace
+instance HasSeverityAnnotation NtN.AcceptConnectionsPolicyTrace where
+  getSeverityAnnotation NtN.ServerTraceAcceptConnectionRateLimiting {} = Info
+  getSeverityAnnotation NtN.ServerTraceAcceptConnectionHardLimit {} = Warning
+
+
+instance HasPrivacyAnnotation (TraceFetchClientState header)
+instance HasSeverityAnnotation (TraceFetchClientState header) where
+  getSeverityAnnotation _ = Info
+
+
+instance HasPrivacyAnnotation (TraceSendRecv a)
+instance HasSeverityAnnotation (TraceSendRecv a) where
+  getSeverityAnnotation _ = Debug
+
+
+instance HasPrivacyAnnotation a => HasPrivacyAnnotation (TraceLabelPeer peer a)
+instance HasSeverityAnnotation a => HasSeverityAnnotation (TraceLabelPeer peer a)
+
+
+instance HasPrivacyAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]
+instance HasSeverityAnnotation [TraceLabelPeer peer (FetchDecision [Point header])] where
+  getSeverityAnnotation =
+      maximum
+    . map (\(TraceLabelPeer _ a) -> fetchDecisionSeverity a)
+    where
+      fetchDecisionSeverity :: FetchDecision a -> Severity
+      fetchDecisionSeverity fd =
+        case fd of
+          Left FetchDeclineChainNotPlausible     -> Debug
+          Left FetchDeclineChainNoIntersection   -> Notice
+          Left FetchDeclineAlreadyFetched        -> Debug
+          Left FetchDeclineInFlightThisPeer      -> Debug
+          Left FetchDeclineInFlightOtherPeer     -> Debug
+          Left FetchDeclinePeerShutdown          -> Info
+          Left FetchDeclinePeerSlow              -> Info
+          Left FetchDeclineReqsInFlightLimit {}  -> Info
+          Left FetchDeclineBytesInFlightLimit {} -> Info
+          Left FetchDeclinePeerBusy {}           -> Info
+          Left FetchDeclineConcurrencyLimit {}   -> Info
+          Right _                                -> Info
+
+
+instance HasPrivacyAnnotation (TraceTxSubmissionInbound txid tx)
+instance HasSeverityAnnotation (TraceTxSubmissionInbound txid tx) where
+  getSeverityAnnotation _ = Info
+
+
+instance HasPrivacyAnnotation (TraceTxSubmissionOutbound txid tx)
+instance HasSeverityAnnotation (TraceTxSubmissionOutbound txid tx) where
+  getSeverityAnnotation _ = Info
+
+
+instance HasPrivacyAnnotation (WithAddr addr ErrorPolicyTrace)
+instance HasSeverityAnnotation (WithAddr addr ErrorPolicyTrace) where
+  getSeverityAnnotation (WithAddr _ ev) = case ev of
+    ErrorPolicySuspendPeer {} -> Warning -- peer misbehaved
+    ErrorPolicySuspendConsumer {} -> Notice -- peer temporarily not useful
+    ErrorPolicyLocalNodeError {} -> Error
+    ErrorPolicyResumePeer {} -> Debug
+    ErrorPolicyKeepSuspended {} -> Debug
+    ErrorPolicyResumeConsumer {} -> Debug
+    ErrorPolicyResumeProducer {} -> Debug
+    ErrorPolicyUnhandledApplicationException {} -> Error
+    ErrorPolicyUnhandledConnectionException {} -> Error
+    ErrorPolicyAcceptException {} -> Error
+
+
+instance HasPrivacyAnnotation (WithDomainName DnsTrace)
+instance HasSeverityAnnotation (WithDomainName DnsTrace) where
+  getSeverityAnnotation (WithDomainName _ ev) = case ev of
+    DnsTraceLookupException {} -> Error
+    DnsTraceLookupAError {} -> Error
+    DnsTraceLookupAAAAError {} -> Error
+    DnsTraceLookupIPv6First -> Debug
+    DnsTraceLookupIPv4First -> Debug
+    DnsTraceLookupAResult {} -> Debug
+    DnsTraceLookupAAAAResult {} -> Debug
+
+
+instance HasPrivacyAnnotation (WithDomainName (SubscriptionTrace Socket.SockAddr))
+instance HasSeverityAnnotation (WithDomainName (SubscriptionTrace Socket.SockAddr)) where
+  getSeverityAnnotation (WithDomainName _ ev) = case ev of
+    SubscriptionTraceConnectStart {} -> Notice
+    SubscriptionTraceConnectEnd {} -> Notice
+    SubscriptionTraceConnectException _ e ->
+        case fromException $ SomeException e of
+             Just (_::SubscriberError) -> Debug
+             Nothing -> Error
+    SubscriptionTraceSocketAllocationException {} -> Error
+    SubscriptionTraceTryConnectToPeer {} -> Info
+    SubscriptionTraceSkippingPeer {} -> Info
+    SubscriptionTraceSubscriptionRunning -> Debug
+    SubscriptionTraceSubscriptionWaiting {} -> Debug
+    SubscriptionTraceSubscriptionFailed -> Warning
+    SubscriptionTraceSubscriptionWaitingNewConnection {} -> Debug
+    SubscriptionTraceStart {} -> Debug
+    SubscriptionTraceRestart {} -> Debug
+    SubscriptionTraceConnectionExist {} -> Info
+    SubscriptionTraceUnsupportedRemoteAddr {} -> Warning
+    SubscriptionTraceMissingLocalAddress -> Warning
+    SubscriptionTraceApplicationException _ e ->
+        case fromException $ SomeException e of
+             Just (_::SubscriberError) -> Debug
+             Nothing -> Error
+    SubscriptionTraceAllocateSocket {} -> Debug
+    SubscriptionTraceCloseSocket {} -> Debug
+
+
+instance HasPrivacyAnnotation (WithIPList (SubscriptionTrace Socket.SockAddr))
+instance HasSeverityAnnotation (WithIPList (SubscriptionTrace Socket.SockAddr)) where
+  getSeverityAnnotation (WithIPList _ _ ev) = case ev of
+    SubscriptionTraceConnectStart _ -> Info
+    SubscriptionTraceConnectEnd _ connectResult -> case connectResult of
+      ConnectSuccess -> Info
+      ConnectSuccessLast -> Notice
+      ConnectValencyExceeded -> Warning
+    SubscriptionTraceConnectException _ e ->
+        case fromException $ SomeException e of
+             Just (_::SubscriberError) -> Debug
+             Nothing -> Error
+    SubscriptionTraceSocketAllocationException {} -> Error
+    SubscriptionTraceTryConnectToPeer {} -> Info
+    SubscriptionTraceSkippingPeer {} -> Info
+    SubscriptionTraceSubscriptionRunning -> Debug
+    SubscriptionTraceSubscriptionWaiting {} -> Debug
+    SubscriptionTraceSubscriptionFailed -> Error
+    SubscriptionTraceSubscriptionWaitingNewConnection {} -> Notice
+    SubscriptionTraceStart {} -> Debug
+    SubscriptionTraceRestart {} -> Info
+    SubscriptionTraceConnectionExist {} -> Notice
+    SubscriptionTraceUnsupportedRemoteAddr {} -> Error
+    SubscriptionTraceMissingLocalAddress -> Warning
+    SubscriptionTraceApplicationException _ e ->
+        case fromException $ SomeException e of
+             Just (_::SubscriberError) -> Debug
+             Nothing -> Error
+    SubscriptionTraceAllocateSocket {} -> Debug
+    SubscriptionTraceCloseSocket {} -> Info
+
+
+instance HasPrivacyAnnotation (Identity (SubscriptionTrace LocalAddress))
+instance HasSeverityAnnotation (Identity (SubscriptionTrace LocalAddress)) where
+  getSeverityAnnotation (Identity ev) = case ev of
+    SubscriptionTraceConnectStart {} -> Notice
+    SubscriptionTraceConnectEnd {} -> Notice
+    SubscriptionTraceConnectException {} -> Error
+    SubscriptionTraceSocketAllocationException {} -> Error
+    SubscriptionTraceTryConnectToPeer {} -> Notice
+    SubscriptionTraceSkippingPeer {} -> Info
+    SubscriptionTraceSubscriptionRunning -> Notice
+    SubscriptionTraceSubscriptionWaiting {} -> Debug
+    SubscriptionTraceSubscriptionFailed -> Warning
+    SubscriptionTraceSubscriptionWaitingNewConnection {} -> Debug
+    SubscriptionTraceStart {} -> Notice
+    SubscriptionTraceRestart {} -> Notice
+    SubscriptionTraceConnectionExist {} -> Debug
+    SubscriptionTraceUnsupportedRemoteAddr {} -> Warning
+    SubscriptionTraceMissingLocalAddress -> Warning
+    SubscriptionTraceApplicationException {} -> Error
+    SubscriptionTraceAllocateSocket {} -> Debug
+    SubscriptionTraceCloseSocket {} -> Debug
+
+
+instance Transformable Text IO (Identity (SubscriptionTrace LocalAddress)) where
+  trTransformer = trStructuredText
+instance HasTextFormatter (Identity (SubscriptionTrace LocalAddress)) where
+  formatText _ = pack . show . toList
+
+
+instance ToObject (Identity (SubscriptionTrace LocalAddress)) where
+  toObject _verb (Identity ev) =
+    mkObject [ "kind" .= ("SubscriptionTrace" :: String)
+             , "event" .= show ev
+             ]
+
+
+instance HasPrivacyAnnotation (WithMuxBearer peer MuxTrace)
+instance HasSeverityAnnotation (WithMuxBearer peer MuxTrace) where
+  getSeverityAnnotation (WithMuxBearer _ ev) = case ev of
+    MuxTraceRecvHeaderStart -> Debug
+    MuxTraceRecvHeaderEnd {} -> Debug
+    MuxTraceRecvStart {} -> Debug
+    MuxTraceRecvEnd {} -> Debug
+    MuxTraceSendStart {} -> Debug
+    MuxTraceSendEnd -> Debug
+    MuxTraceState {} -> Info
+    MuxTraceCleanExit {} -> Notice
+    MuxTraceExceptionExit {} -> Notice
+    MuxTraceChannelRecvStart {} -> Debug
+    MuxTraceChannelRecvEnd {} -> Debug
+    MuxTraceChannelSendStart {} -> Debug
+    MuxTraceChannelSendEnd {} -> Debug
+    MuxTraceHandshakeStart -> Debug
+    MuxTraceHandshakeClientEnd {} -> Info
+    MuxTraceHandshakeServerEnd -> Debug
+    MuxTraceHandshakeClientError {} -> Error
+    MuxTraceHandshakeServerError {} -> Error
+    MuxTraceRecvDeltaQObservation {} -> Debug
+    MuxTraceRecvDeltaQSample {} -> Debug
+    MuxTraceSDUReadTimeoutException -> Notice
+
+--
+-- | instances of @Transformable@
+--
+-- NOTE: this list is sorted by the unqualified name of the outermost type.
+
+instance Transformable Text IO NtN.HandshakeTr where
+  trTransformer = trStructuredText
+instance HasTextFormatter NtN.HandshakeTr where
+  formatText _ = pack . show . toList
+
+
+instance Transformable Text IO NtC.HandshakeTr where
+  trTransformer = trStructuredText
+instance HasTextFormatter NtC.HandshakeTr where
+  formatText _ = pack . show . toList
+
+
+instance Transformable Text IO NtN.AcceptConnectionsPolicyTrace where
+  trTransformer = trStructuredText
+instance HasTextFormatter NtN.AcceptConnectionsPolicyTrace where
+  formatText _ = pack . show . toList
+
+
+instance Show peer
+      => Transformable Text IO [TraceLabelPeer peer (FetchDecision [Point header])] where
+  trTransformer = trStructuredText
+instance HasTextFormatter [TraceLabelPeer peer (FetchDecision [Point header])] where
+  formatText _ = pack . show . toList
+
+
+instance (Show peer, HasPrivacyAnnotation a, HasSeverityAnnotation a, ToObject a)
+      => Transformable Text IO (TraceLabelPeer peer a) where
+  trTransformer = trStructuredText
+instance HasTextFormatter (TraceLabelPeer peer a) where
+  formatText _ = pack . show . toList
+
+
+instance Transformable Text IO (TraceTxSubmissionInbound txid tx) where
+  trTransformer = trStructuredText
+instance HasTextFormatter (TraceTxSubmissionInbound txid tx) where
+  formatText _ = pack . show . toList
+
+
+instance (Show tx, Show txid)
+      => Transformable Text IO (TraceTxSubmissionOutbound txid tx) where
+  trTransformer = trStructuredText
+instance HasTextFormatter (TraceTxSubmissionOutbound txid tx) where
+  formatText _ = pack . show . toList
+
+
+instance Show addr => Transformable Text IO (WithAddr addr ErrorPolicyTrace) where
+  trTransformer = trStructuredText
+instance HasTextFormatter (WithAddr addr ErrorPolicyTrace) where
+  formatText _ = pack . show . toList
+
+
+instance Transformable Text IO (WithDomainName (SubscriptionTrace Socket.SockAddr)) where
+  trTransformer = trStructuredText
+instance HasTextFormatter (WithDomainName (SubscriptionTrace Socket.SockAddr)) where
+  formatText _ = pack . show . toList
+
+
+instance Transformable Text IO (WithDomainName DnsTrace) where
+  trTransformer = trStructuredText
+instance HasTextFormatter (WithDomainName DnsTrace) where
+  formatText _ = pack . show . toList
+
+
+instance Transformable Text IO (WithIPList (SubscriptionTrace Socket.SockAddr)) where
+  trTransformer = trStructuredText
+instance HasTextFormatter (WithIPList (SubscriptionTrace Socket.SockAddr)) where
+  formatText _ = pack . show . toList
+
+
+instance (Show peer)
+      => Transformable Text IO (WithMuxBearer peer MuxTrace) where
+  trTransformer = trStructuredText
+instance (Show peer)
+      => HasTextFormatter (WithMuxBearer peer MuxTrace) where
+  formatText (WithMuxBearer peer ev) = \_o ->
+    "Bearer on " <> pack (show peer)
+   <> " event: " <> pack (show ev)
+
+
+--
+-- | instances of @ToObject@
+--
+-- NOTE: this list is sorted by the unqualified name of the outermost type.
+
+instance ( Condense (HeaderHash blk)
+         , Condense (TxId (GenTx blk))
+         , HasHeader blk
+         , HasTxs blk
+         , HasTxId (GenTx blk)
+         )
+      => ToObject (AnyMessage (BlockFetch blk)) where
+  toObject _verb (AnyMessage (MsgBlock blk)) =
+    mkObject
+      [ "kind" .= String "MsgBlock"
+      , "blkid" .= String (pack . condense $ blockHash blk)
+      , "txids" .= toJSON (presentTx <$> extractTxs blk)
+      ]
+   where
+     presentTx :: GenTx blk -> Value
+     presentTx =  String . pack . condense . txId
+  toObject _v (AnyMessage MsgRequestRange{}) =
+    mkObject [ "kind" .= String "MsgRequestRange" ]
+  toObject _v (AnyMessage MsgStartBatch{}) =
+    mkObject [ "kind" .= String "MsgStartBatch" ]
+  toObject _v (AnyMessage MsgNoBlocks{}) =
+    mkObject [ "kind" .= String "MsgNoBlocks" ]
+  toObject _v (AnyMessage MsgBatchDone{}) =
+    mkObject [ "kind" .= String "MsgBatchDone" ]
+  toObject _v (AnyMessage MsgClientDone{}) =
+    mkObject [ "kind" .= String "MsgClientDone" ]
+
+
+instance ToObject (FetchDecision [Point header]) where
+  toObject _verb (Left decline) =
+    mkObject [ "kind" .= String "FetchDecision declined"
+             , "declined" .= String (pack $ show $ decline) ]
+  toObject _verb (Right results) =
+    mkObject [ "kind" .= String "FetchDecision results"
+             , "length" .= String (pack $ show $ length results) ]
+
+
+instance ToObject NtC.HandshakeTr where
+  toObject _verb (WithMuxBearer b ev) =
+    mkObject [ "kind" .= String "LocalHandshakeTrace"
+             , "bearer" .= show b
+             , "event" .= show ev ]
+
+
+instance ToObject NtN.HandshakeTr where
+  toObject _verb (WithMuxBearer b ev) =
+    mkObject [ "kind" .= String "HandshakeTrace"
+             , "bearer" .= show b
+             , "event" .= show ev ]
+
+
+instance ToObject NtN.AcceptConnectionsPolicyTrace where
+  toObject _verb (NtN.ServerTraceAcceptConnectionRateLimiting delay numOfConnections) =
+    mkObject [ "kind" .= String "ServerTraceAcceptConnectionRateLimiting"
+             , "delay" .= show delay
+             , "numberOfConnection" .= show numOfConnections
+             ]
+  toObject _verb (NtN.ServerTraceAcceptConnectionHardLimit softLimit) =
+    mkObject [ "kind" .= String "ServerTraceAcceptConnectionHardLimit"
+             , "softLimit" .= show softLimit
+             ]
+
+
+instance (Show txid, Show tx)
+      => ToObject (AnyMessage (TxSubmission txid tx)) where
+  toObject _verb (AnyMessage (MsgRequestTxs txids)) =
+    mkObject
+      [ "kind" .= String "MsgRequestTxs"
+      , "txIds" .= String (pack $ show txids)
+      ]
+  toObject _verb (AnyMessage (MsgReplyTxs txs)) =
+    mkObject
+      [ "kind" .= String "MsgReplyTxs"
+      , "txs" .= String (pack $ show txs)
+      ]
+  toObject _verb (AnyMessage (MsgRequestTxIds _ _ _)) =
+    mkObject
+      [ "kind" .= String "MsgRequestTxIds"
+      ]
+  toObject _verb (AnyMessage (MsgReplyTxIds _)) =
+    mkObject
+      [ "kind" .= String "MsgReplyTxIds"
+      ]
+  toObject _verb (AnyMessage MsgDone) =
+    mkObject
+      [ "kind" .= String "MsgDone"
+      ]
+
+
+instance Condense (HeaderHash blk)
+      => ToObject (Point blk) where
+  toObject MinimalVerbosity p = toObject NormalVerbosity p
+  toObject verb p =
+    mkObject [ "kind" .= String "Tip" --TODO: why is this a Tip not a Point?
+             , "tip" .= showPoint verb p ]
+
+
+instance ToObject SlotNo where
+  toObject _verb slot =
+    mkObject [ "kind" .= String "SlotNo"
+             , "slot" .= toJSON (unSlotNo slot) ]
+
+
+instance ToObject (TraceFetchClientState header) where
+  toObject _verb (AddedFetchRequest {}) =
+    mkObject [ "kind" .= String "AddedFetchRequest" ]
+  toObject _verb (AcknowledgedFetchRequest {}) =
+    mkObject [ "kind" .= String "AcknowledgedFetchRequest" ]
+  toObject _verb (CompletedBlockFetch {}) =
+    mkObject [ "kind" .= String "CompletedBlockFetch" ]
+  toObject _verb (CompletedFetchBatch {}) =
+    mkObject [ "kind" .= String "CompletedFetchBatch" ]
+  toObject _verb (StartedFetchBatch {}) =
+    mkObject [ "kind" .= String "StartedFetchBatch" ]
+  toObject _verb (RejectedFetchBatch {}) =
+    mkObject [ "kind" .= String "RejectedFetchBatch" ]
+
+
+instance Show peer
+      => ToObject [TraceLabelPeer peer (FetchDecision [Point header])] where
+  toObject MinimalVerbosity _ = emptyObject
+  toObject _ [] = emptyObject
+  toObject _ (lbl : r) = toObject MaximalVerbosity lbl <>
+                                        toObject MaximalVerbosity r
+
+
+instance (Show peer, ToObject a) => ToObject (TraceLabelPeer peer a) where
+  toObject verb (TraceLabelPeer peerid a) =
+    mkObject [ "peer" .= show peerid ] <> toObject verb a
+
+
+instance ToObject (AnyMessage ps)
+      => ToObject (TraceSendRecv ps) where
+  toObject verb (TraceSendMsg m) = mkObject
+    [ "kind" .= String "Send" , "msg" .= toObject verb m ]
+  toObject verb (TraceRecvMsg m) = mkObject
+    [ "kind" .= String "Recv" , "msg" .= toObject verb m ]
+
+
+instance ToObject (TraceTxSubmissionInbound txid tx) where
+  toObject _verb TraceTxSubmissionInbound =
+    mkObject [ "kind" .= String "TraceTxSubmissionInbound" ]
+
+
+instance (Show txid, Show tx)
+      => ToObject (TraceTxSubmissionOutbound txid tx) where
+  toObject MaximalVerbosity (TraceTxSubmissionOutboundRecvMsgRequestTxs txids) =
+    mkObject
+      [ "kind" .= String "TraceTxSubmissionOutboundRecvMsgRequestTxs"
+      , "txIds" .= String (pack $ show txids)
+      ]
+  toObject _verb (TraceTxSubmissionOutboundRecvMsgRequestTxs _txids) =
+    mkObject
+      [ "kind" .= String "TraceTxSubmissionOutboundRecvMsgRequestTxs"
+      ]
+  toObject MaximalVerbosity (TraceTxSubmissionOutboundSendMsgReplyTxs txs) =
+    mkObject
+      [ "kind" .= String "TraceTxSubmissionOutboundSendMsgReplyTxs"
+      , "txs" .= String (pack $ show txs)
+      ]
+  toObject _verb (TraceTxSubmissionOutboundSendMsgReplyTxs _txs) =
+    mkObject
+      [ "kind" .= String "TraceTxSubmissionOutboundSendMsgReplyTxs"
+      ]
+
+instance Show addr => ToObject (WithAddr addr ErrorPolicyTrace) where
+  toObject _verb (WithAddr addr ev) =
+    mkObject [ "kind" .= String "ErrorPolicyTrace"
+             , "address" .= show addr
+             , "event" .= show ev ]
+
+
+instance ToObject (WithIPList (SubscriptionTrace Socket.SockAddr)) where
+  toObject _verb (WithIPList localAddresses dests ev) =
+    mkObject [ "kind" .= String "WithIPList SubscriptionTrace"
+             , "localAddresses" .= show localAddresses
+             , "dests" .= show dests
+             , "event" .= show ev ]
+
+
+instance ToObject (WithDomainName DnsTrace) where
+  toObject _verb (WithDomainName dom ev) =
+    mkObject [ "kind" .= String "DnsTrace"
+             , "domain" .= show dom
+             , "event" .= show ev ]
+
+
+instance ToObject (WithDomainName (SubscriptionTrace Socket.SockAddr)) where
+  toObject _verb (WithDomainName dom ev) =
+    mkObject [ "kind" .= String "SubscriptionTrace"
+             , "domain" .= show dom
+             , "event" .= show ev ]
+
+
+instance (Show peer) => ToObject (WithMuxBearer peer MuxTrace) where
+  toObject _verb (WithMuxBearer b ev) =
+    mkObject [ "kind" .= String "MuxTrace"
+             , "bearer" .= show b
+             , "event" .= show ev ]
