+++ cardano-node/src/Cardano/Tracing/Tracers.hs
@@ -69,7 +69,6 @@ import qualified Ouroboros.Consensus.Storage.LedgerDB.OnDisk as LedgerDB
 import           Cardano.Config.Protocol (TraceConstraints)
 import           Cardano.Config.TraceConfig
 import           Cardano.Tracing.MicroBenchmarking
-import           Cardano.Tracing.ToObjectOrphans ()
 
 import           Control.Tracer.Transformers
